Move CJITIncludeLoader and make it work nicely!
