+++ include/nbl/video/CJITIncludeLoader.h
@@ -15,11 +15,15 @@ namespace nbl::video
 class NBL_API2 CJITIncludeLoader : public asset::IShaderCompiler::IIncludeLoader
 {
     public:
-        CJITIncludeLoader(const SPhysicalDeviceLimits& limits, const SPhysicalDeviceFeatures& features);
-        std::optional<std::string> getInclude(const system::path& searchPath, const std::string& includeName) const override;
+        inline CJITIncludeLoader(const SPhysicalDeviceLimits& limits, const SPhysicalDeviceFeatures& features)
+        {
+            m_includes["nbl/builtin/hlsl/jit/device_capabilities.hlsl"] = collectDeviceCaps(limits,features);
+        }
+
+        found_t getInclude(const system::path& searchPath, const std::string& includeName) const override;
 
     private:
-        core::unordered_map<system::path, std::string> m_includes;
+        core::unordered_map<system::path,std::string> m_includes;
         std::string collectDeviceCaps(const SPhysicalDeviceLimits& limits, const SPhysicalDeviceFeatures& features);
 
 };
